Initial Upload
This upload includes the Map for Lonestar Station (Vima Facility) and site specific roles for players.

Known issues include all of this not working properly due to: A lack of an engine, a lack of properly placed airlocks, a lack of properly placed cameras, web shuttle system not set up for local shuttles, questionable power grid, no cowboy boots, role selection not currently functioning, arrivals not arriving, latejoin not joining, evacuation not evacuating, and most if not all departments are missing key features and equipment.